% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UESI_ggtheme.R
\name{format_city_name}
\alias{format_city_name}
\title{Function for formatting the city names}
\usage{
format_city_name(x)
}
\arguments{
\item{x}{A vector of citynames}
}
\value{
Formatted city names ready for plotting
}
\description{
Use this function to format City names for plots within the UESI project
}
