% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UESI_ggtheme.R
\name{theme_UESI}
\alias{theme_UESI}
\title{UESI ggplot theme function}
\usage{
theme_UESI()
}
\description{
Standardized theme for ggplots for use in the UESI project. The function follows similar
usage as the "theme" function in ggplot2.
}
\keyword{UESI}
\keyword{ggplot}
